chore: make compatible with latest Homebrew version
Homebrew removed the `brew cask` command in favor of
unifying the commands for formulae and casks while
providing options to specify which of these you want
(e.g., using `--formula` or `--cask`).

Additionally, `brew info --json=v2` is a new version of
the JSON-based info command which now includes support for
casks, too.

This allowed for unifying the functions used for both cases.

Naming collisions between formulae (homebrew/core) and casks (homebrew/cask)
can exist and in these cases, Homebrew will pick the formula and print a warning.
It is possible to use a qualified identifier to point to a specific one.

This was necessary i.e. for "docker" as we have the cask for the Docker CE desktop app
as well as the CLI(-only) formula.

A conflict at the upgrade procedure at miniconda
required to change the order and install/upgrade Python first.

Fixes issue #8.