feat: remove intellij-idea managed via brew
Installing the IDE via Homebrew *and* the Toolbox can cause some trouble,
esp. for upgrades. The IDE might be installed multiple times, etc.

Therefore, the IDE will need to be installed manually for now using the Toolbox.
Upgrades, etc. can be managed by the Toolbox (manually or automatically).

To be investigated whether we can trigger the installation for desired tools
via Toolbox from the CLI.

relates to issue #14